lots o stuff

+++ db/migrate/20241220212131_create_batches.rb
@@ -3,7 +3,7 @@ class CreateBatches < ActiveRecord::Migration[7.2]
     create_table :batches do |t|
       t.string :name, null: false
       t.date :manufactured_on
-      t.decimal :units_produced
+      t.integer :units_produced
 
       t.timestamps
     end
